feat: refactor hooks for l1 messenger and burn native token
