+++ tests/instances/transactions/Cargo.toml
@@ -11,6 +11,7 @@ categories.workspace = true
 
 [dependencies]
 rig = { path = "../../rig", features = ["for_tests"] }
+system_hooks = { path = "../../../system_hooks" }
 hex = "*"
 
 [features]
