+++ tests/instances/transactions/src/lib.rs
@@ -14,8 +14,10 @@ use rig::zksync_os_interface::error::InvalidTransaction;
 use rig::{alloy, ethers, zksync_web3_rs, Chain};
 use rig::{utils::*, BlockContext};
 use std::str::FromStr;
+use system_hooks::addresses_constants::{L1_MESSENGER_ADDRESS, L2_BASE_TOKEN_ADDRESS};
 use zksync_web3_rs::eip712::Eip712Meta;
 use zksync_web3_rs::signers::{LocalWallet, Signer};
+
 mod native_charging;
 
 fn run_config() -> Option<rig::chain::RunConfig> {
@@ -275,6 +277,9 @@ fn test_gas_price_zero() {
 #[test]
 fn test_withdrawal() {
     let mut chain = Chain::empty(None);
+    let bytecode_vec: Vec<u8> = hex::decode("60806040526004361015610013575b610129565b61001e60003561002d565b6362f84b240361000e576100f3565b60e01c90565b60405190565b600080fd5b600080fd5b600080fd5b600080fd5b600080fd5b600080fd5b909182601f830112156100915781359167ffffffffffffffff831161008c57602001926001830284011161008757565b610052565b61004d565b610048565b906020828203126100c857600082013567ffffffffffffffff81116100c3576100bf9201610057565b9091565b610043565b61003e565b90565b6100d9906100cd565b9052565b91906100f1906000602085019401906100d0565b565b346101245761012061010f610109366004610096565b90610404565b610117610033565b918291826100dd565b0390f35b610039565b600080fd5b600090565b90565b60018060a01b031690565b90565b61015861015361015d92610133565b610141565b610136565b90565b61016b617000610144565b90565b90565b61018561018061018a9261016e565b610141565b610136565b90565b634e487b7160e01b600052601160045260246000fd5b6101af6101b591610136565b91610136565b019060018060a01b0382116101c657565b61018d565b6101df6101da6101e492610136565b610141565b610136565b90565b6101f0906101cb565b90565b610216610211610201610160565b61020b6001610171565b906101a3565b6101e7565b90565b905090565b90826000939282370152565b90918261023a8161024193610219565b809361021e565b0190565b90916102509261022a565b90565b601f801991011690565b634e487b7160e01b600052604160045260246000fd5b9061027d90610253565b810190811067ffffffffffffffff82111761029757604052565b61025d565b906102af6102a8610033565b9283610273565b565b67ffffffffffffffff81116102cf576102cb602091610253565b0190565b61025d565b906102e66102e1836102b1565b61029c565b918252565b606090565b3d60001461030d576103013d6102d4565b903d6000602084013e5b565b6103156102eb565b9061030b565b151590565b5190565b90565b90565b61033e61033961034392610327565b610141565b610324565b90565b60000190565b60200190565b61035c90516100cd565b90565b1b90565b61037d61037861037283610320565b9261034c565b610352565b906020811061038b575b5090565b61039e906000199060200360080261035f565b1638610387565b6103ae906101e7565b90565b6103ba906100cd565b90565b60209181520190565b91906103e0816103d9816103e5956103bd565b809561021e565b610253565b0190565b909161040192602083019260008185039101526103c6565b90565b919061040e61012e565b5060008061041a6101f3565b85828591610432610429610033565b93849283610245565b03925af16104486104416102f0565b911561031b565b80156104cb575b6104ae5761045c90610363565b923384919261049461048e7f3a36e47291f4201faf137fab081d92295bce2d53be2c6ca68ba82c7faa9ce241936103a5565b936103b1565b936104a96104a0610033565b928392836103e9565b0390a3565b600063f801b06960e01b8152806104c760048201610346565b0390fd5b506104d581610320565b6104e86104e2602061032a565b91610324565b141561044f56fea2646970667358221220eaa72a072e95715690c88c25da2fd94b2b0a7a610c93721e3eb39b3c9804086464736f6c634300081c0033").expect("valid hex");
+    let l1_messenger_hook_bytecode: &[u8] = &bytecode_vec; // borrow as slice
+    chain.set_evm_bytecode(L1_MESSENGER_ADDRESS, l1_messenger_hook_bytecode);
 
     let wallet = PrivateKeySigner::from_str(
         "dcf2cbdd171a21c480aa7f53d77f31bb102282b3ff099c78e3118b37348c72f7",
@@ -336,6 +341,10 @@ fn test_withdrawal() {
         B160::from_be_bytes(from.0),
         U256::from(1_000_000_000_000_000_u64),
     );
+
+    let bytecode_vec_l2_base_token: Vec<u8> = hex::decode("60806040526004361015610013575b61019b565b61001e60003561003d565b806351cff8d914610038576384bc3eb00361000e57610171565b6100b3565b60e01c90565b60405190565b600080fd5b600080fd5b60018060a01b031690565b61006790610053565b90565b6100738161005e565b0361007a57565b600080fd5b9050359061008c8261006a565b565b906020828203126100a8576100a59160000161007f565b90565b610049565b60000190565b6100c66100c136600461008e565b6103fb565b6100ce610043565b806100d8816100ad565b0390f35b600080fd5b600080fd5b600080fd5b909182601f830112156101255781359167ffffffffffffffff831161012057602001926001830284011161011b57565b6100e6565b6100e1565b6100dc565b91909160408184031261016c57610144836000830161007f565b92602082013567ffffffffffffffff81116101675761016392016100eb565b9091565b61004e565b610049565b61018561017f36600461012a565b916105b7565b61018d610043565b80610197816100ad565b0390f35b600080fd5b90565b90565b6101ba6101b56101bf926101a0565b6101a3565b610053565b90565b6101cd6180006101a6565b90565b90565b6101e76101e26101ec926101d0565b6101a3565b610053565b90565b634e487b7160e01b600052601160045260246000fd5b61021161021791610053565b91610053565b019060018060a01b03821161022857565b6101ef565b61024161023c61024692610053565b6101a3565b610053565b90565b6102529061022d565b90565b61025e9061022d565b90565b61026a90610255565b90565b61029861029361028e61027e6101c2565b61028860086101d3565b90610205565b610249565b610261565b90565b6102a490610249565b90565b601f801991011690565b634e487b7160e01b600052604160045260246000fd5b906102d1906102a7565b810190811067ffffffffffffffff8211176102eb57604052565b6102b1565b60e01b90565b90565b610302816102f6565b0361030957565b600080fd5b9050519061031b826102f9565b565b90602082820312610337576103349160000161030e565b90565b610049565b5190565b60209181520190565b60005b83811061035d575050906000910152565b80602091830151818501520161034c565b61038d61039660209361039b936103848161033c565b93848093610340565b95869101610349565b6102a7565b0190565b6103b5916020820191600081840391015261036e565b90565b6103c0610043565b3d6000823e3d90fd5b6103d290610249565b90565b90565b6103e1906103d5565b9052565b91906103f9906000602085019401906103d8565b565b6104036107a7565b61044c60206104138484906108e4565b61042361041e61026d565b61029b565b61044160006362f84b24610435610043565b968795869485936102f0565b83526004830161039f565b03925af180156104d9576104ad575b50339190916104a86104966104907f2717ead6b9200dd235aad468c9809ea400fe33ac69b5bfaa6d3e90fc922b6398936103c9565b936103c9565b9361049f610043565b918291826103e5565b0390a3565b6104cd9060203d81116104d2575b6104c581836102c7565b81019061031d565b61045b565b503d6104bb565b6103b8565b600080fd5b906104f66104ef610043565b92836102c7565b565b67ffffffffffffffff8111610516576105126020916102a7565b0190565b6102b1565b90826000939282370152565b9092919261053c610537826104f8565b6104e3565b93818552602085019082840111610558576105569261051b565b565b6104de565b610568913691610527565b90565b91906105858161057e8161058a95610340565b809561051b565b6102a7565b0190565b916105b49391926105a7604082019460008301906103d8565b602081850391015261056b565b90565b9190916105c26107a7565b9161061960206105e0848633906105da8a889061055d565b92610991565b6105f06105eb61026d565b61029b565b61060e60006362f84b24610602610043565b968795869485936102f0565b83526004830161039f565b03925af180156106a85761067c575b5033919293909361067761066561065f7fc405fe8958410bbaf0c73b7a0c3e20859e86ca168a4c9b0def9c54d2555a306b956103c9565b956103c9565b9561066e610043565b9384938461058e565b0390a3565b61069c9060203d81116106a1575b61069481836102c7565b81019061031d565b610628565b503d61068a565b6103b8565b600090565b90565b6106c96106c46106ce926106b2565b6101a3565b610053565b90565b6106f46106ef6106df6101c2565b6106e960026106b5565b90610205565b610249565b90565b6107009061022d565b90565b61070c906106f7565b90565b61071890610249565b90565b905090565b61072c6000809261071b565b0190565b61073990610720565b90565b9061074e610749836104f8565b6104e3565b918252565b606090565b3d600014610775576107693d61073c565b903d6000602084013e5b565b61077d610753565b90610773565b151590565b90565b61079f61079a6107a492610788565b6101a3565b6103d5565b90565b6107af6106ad565b5034906000806107cd6107c86107c36106d1565b610703565b61070f565b846107d6610043565b90816107e181610730565b03925af16107f76107f0610758565b9115610783565b908115610823575b5061080657565b6000631bc5aabf60e21b81528061081f600482016100ad565b0390fd5b61082d915061033c565b61084061083a600061078b565b916103d5565b1415386107ff565b63ffffffff60e01b1690565b90565b61086361086891610848565b610854565b9052565b60601b90565b61087b9061086c565b90565b61088790610872565b90565b61089661089b9161005e565b61087e565b9052565b90565b6108ae6108b3916103d5565b61089f565b9052565b602093926108d86004836108d06014956108e097610857565b01809261088a565b0180926108a2565b0190565b610921906108f0610753565b50610912636c0960f960e01b9193610906610043565b948593602085016108b7565b602082018103825203826102c7565b90565b610949610940926020926109378161033c565b9485809361071b565b93849101610349565b0190565b60149361097f8561098e98979561097760048661096f6020986109879a610857565b01809261088a565b0180926108a2565b01809261088a565b0190610924565b90565b90926109d4926109c5916109a3610753565b50636c0960f960e01b93959190916109b9610043565b9687956020870161094d565b602082018103825203826102c7565b9056fea26469706673582212201e2659874946c1e4eedce28cd876068bf3e5496c34be61dc6ed85365e70db29664736f6c634300081c0033").expect("valid hex");
+    let l2_base_token_contract_bytecode: &[u8] = &bytecode_vec_l2_base_token; // borrow as slice
+    chain.set_evm_bytecode(L2_BASE_TOKEN_ADDRESS, l2_base_token_contract_bytecode);
 
     let output = chain.run_block(transactions, None, run_config());
